+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/MineshaftEndStructure.java
@@ -2,20 +2,33 @@ package com.telepathicgrunt.repurposedstructures.world.structures;
 
 import com.telepathicgrunt.repurposedstructures.RepurposedStructures;
 import com.telepathicgrunt.repurposedstructures.modinit.RSStructureTagMap;
+import com.telepathicgrunt.repurposedstructures.world.structures.pieces.PieceLimitedJigsawManager;
+import net.minecraft.block.BlockState;
 import net.minecraft.util.Direction;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.SharedSeedRandom;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.ChunkPos;
+import net.minecraft.util.math.MutableBoundingBox;
 import net.minecraft.util.math.vector.Vector3f;
+import net.minecraft.util.math.vector.Vector3i;
+import net.minecraft.util.registry.DynamicRegistries;
+import net.minecraft.util.registry.Registry;
+import net.minecraft.world.IBlockReader;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.biome.provider.BiomeProvider;
 import net.minecraft.world.gen.ChunkGenerator;
 import net.minecraft.world.gen.Heightmap;
 import net.minecraft.world.gen.feature.NoFeatureConfig;
 import net.minecraft.world.gen.feature.StructureFeature;
+import net.minecraft.world.gen.feature.structure.MarginedStructureStart;
 import net.minecraft.world.gen.feature.structure.Structure;
+import net.minecraft.world.gen.feature.structure.VillageConfig;
+import net.minecraft.world.gen.feature.template.TemplateManager;
 import net.minecraft.world.gen.settings.StructureSeparationSettings;
 import net.minecraftforge.common.util.Lazy;
+
+import javax.swing.text.html.BlockView;
 
 
 public class MineshaftEndStructure extends MineshaftStructure {
@@ -32,10 +45,6 @@ public class MineshaftEndStructure extends MineshaftStructure {
         boolean superCheck = super.isFeatureChunk(chunkGenerator, biomeSource, seed, chunkRandom, chunkX, chunkZ, biome, chunkPos, featureConfig);
         if(!superCheck)
             return false;
-
-        if(RepurposedStructures.RSMineshaftsConfig.barrensIslandsEndMineshafts.get())
-            return true;
-
 
         //cannot be near end strongholds
         int structureCheckRadius = 6;
@@ -55,25 +64,109 @@ public class MineshaftEndStructure extends MineshaftStructure {
             }
         }
 
-        int minLandHeight = Math.min(chunkGenerator.getGenDepth(), 45);
+        int minThickness = RepurposedStructures.RSMineshaftsConfig.endMineshaftMinIslandThickness.get();
+        if(minThickness == 0)
+            return true;
+
+        BlockPos.Mutable islandTopBottomThickness = new BlockPos.Mutable(Integer.MAX_VALUE, Integer.MIN_VALUE, Integer.MAX_VALUE);
         int xPos = chunkX * 16;
         int zPos = chunkZ * 16;
-        int landHeight = getHeightAt(chunkGenerator, xPos, zPos, Integer.MAX_VALUE);
-        if(landHeight < minLandHeight) return false;
 
-        for(Direction direction : Direction.Plane.HORIZONTAL){
-            Vector3f offsetPos = new Vector3f(direction.getStepX(), direction.getStepY(), direction.getStepZ());
-            offsetPos = new Vector3f(offsetPos.x() * 70f, offsetPos.y() * 70f, offsetPos.z() * 70f);
-            landHeight = getHeightAt(chunkGenerator, xPos + (int)offsetPos.x(), zPos + (int)offsetPos.z(), landHeight);
-            if(landHeight < minLandHeight) return false;
+        analyzeLand(chunkGenerator, xPos, zPos, islandTopBottomThickness);
+        if(islandTopBottomThickness.getZ() < minThickness) return false;
+
+        for(int i = 1; i <= 2; i++) {
+            for (Direction direction : Direction.Plane.HORIZONTAL) {
+                Vector3f offsetPos = new Vector3f(direction.getStepX(), 0, direction.getStepZ());
+                offsetPos = new Vector3f(offsetPos.x() * 30f * i, 0, offsetPos.z() * 30f * i);
+                analyzeLand(chunkGenerator, xPos + (int) offsetPos.x(), zPos + (int) offsetPos.z(), islandTopBottomThickness);
+                if (islandTopBottomThickness.getZ() < minThickness) return false;
+            }
         }
 
         return true;
     }
 
-    private int getHeightAt(ChunkGenerator chunkGenerator, int xPos, int zPos, int landHeight) {
-        landHeight = Math.min(landHeight, chunkGenerator.getFirstOccupiedHeight(xPos, zPos, Heightmap.Type.WORLD_SURFACE_WG));
-        return landHeight;
+    private void analyzeLand(ChunkGenerator chunkGenerator, int xPos, int zPos, BlockPos.Mutable islandTopBottomThickness) {
+        IBlockReader columnOfBlocks = chunkGenerator.getBaseColumn(xPos, zPos);
+        BlockPos.Mutable currentPos = new BlockPos.Mutable(xPos, chunkGenerator.getGenDepth(), zPos);
+        boolean isInIsland = false;
+        while(currentPos.getY() >= 0){
+            BlockState state = columnOfBlocks.getBlockState(currentPos);
+            if(!state.isAir() && !isInIsland){
+                isInIsland = true;
+                int topIslandY = Math.min(currentPos.getY(), islandTopBottomThickness.getX());
+                islandTopBottomThickness.set(topIslandY, islandTopBottomThickness.getY(), islandTopBottomThickness.getZ());
+            }
+            else if(state.isAir() && isInIsland){
+                int bottomIslandY = Math.max(currentPos.getY(), islandTopBottomThickness.getY());
+                islandTopBottomThickness.set(islandTopBottomThickness.getX(), bottomIslandY, islandTopBottomThickness.getZ());
+                break;
+            }
+
+            currentPos.move(Direction.DOWN);
+        }
+        // Never hit land
+        if(!isInIsland){
+            islandTopBottomThickness.set(0, islandTopBottomThickness.getY(), islandTopBottomThickness.getZ());
+        }
+
+        int thickness = islandTopBottomThickness.getX() - islandTopBottomThickness.getY();
+        islandTopBottomThickness.set(islandTopBottomThickness.getX(), islandTopBottomThickness.getY(), thickness);
+    }
+
+
+    @Override
+    public IStartFactory<NoFeatureConfig> getStartFactory() {
+        return MineshaftEndStructure.MainStart::new;
+    }
+
+    public class MainStart extends MarginedStructureStart<NoFeatureConfig> {
+
+        private final ResourceLocation structureID;
+
+        public MainStart(Structure<NoFeatureConfig> structureIn, int chunkX, int chunkZ, MutableBoundingBox box, int referenceIn, long seedIn) {
+            super(structureIn, chunkX, chunkZ, box, referenceIn, seedIn);
+            structureID = Registry.STRUCTURE_FEATURE.getKey(structureIn);
+        }
+
+        public void generatePieces(DynamicRegistries dynamicRegistryManager, ChunkGenerator chunkGenerator, TemplateManager structureManager, int chunkX, int chunkZ, Biome biome, NoFeatureConfig defaultFeatureConfig) {
+            BlockPos.Mutable blockpos = new BlockPos.Mutable(chunkX * 16, 0, chunkZ * 16);
+            BlockPos.Mutable islandTopBottomThickness = new BlockPos.Mutable(Integer.MAX_VALUE, Integer.MIN_VALUE, Integer.MAX_VALUE);
+            analyzeLand(chunkGenerator, blockpos.getX(), blockpos.getZ(), islandTopBottomThickness);
+
+            int minThickness = RepurposedStructures.RSMineshaftsConfig.endMineshaftMinIslandThickness.get();
+            int maxY = 53;
+            int minY = 15;
+            if(minThickness == 0){
+                blockpos.move(Direction.UP, 35);
+            }
+            else{
+                int structureStartHeight = random.nextInt((islandTopBottomThickness.getZ() - minThickness) + 1) + islandTopBottomThickness.getY() + (minThickness / 2);
+                blockpos.move(Direction.UP, structureStartHeight);
+                maxY = islandTopBottomThickness.getX() - 5;
+                minY = islandTopBottomThickness.getY();
+                if(maxY - minY <= 5){
+                    minY = maxY - 5;
+                }
+            }
+
+            PieceLimitedJigsawManager.assembleJigsawStructure(
+                    dynamicRegistryManager,
+                    new VillageConfig(() -> dynamicRegistryManager.registryOrThrow(Registry.TEMPLATE_POOL_REGISTRY).get(startPool), structureSize.get()),
+                    chunkGenerator,
+                    structureManager,
+                    blockpos,
+                    this.pieces,
+                    this.random,
+                    false,
+                    false,
+                    structureID,
+                    maxY,
+                    minY);
+
+            this.calculateBoundingBox();
+        }
     }