+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/StrongholdEndStructure.java
@@ -37,8 +37,8 @@ public class StrongholdEndStructure extends AdvancedDistanceJigsawStructure {
 
         for(int i = 1; i <= 2; i++){
             for(Direction direction : Direction.Plane.HORIZONTAL){
-                Vector3f offsetPos = new Vector3f(direction.getStepX(), direction.getStepY(), direction.getStepZ());
-                offsetPos = new Vector3f(offsetPos.x() * 35f * i, offsetPos.y() * 35f * i, offsetPos.z() * 35f * i);
+                Vector3f offsetPos = new Vector3f(direction.getStepX(), 0, direction.getStepZ());
+                offsetPos = new Vector3f(offsetPos.x() * 35f * i, 0, offsetPos.z() * 35f * i);
                 landHeight = getHeightAt(chunkGenerator, xPos + (int)offsetPos.x(), zPos + (int)offsetPos.z(), landHeight);
                 if(landHeight < minLandHeight) return false;
             }